feat(backend): add api endpoints for blog, featured, and project
Backend finalized. Only the media service remains to be implemented.
For now, we can bypass the media service by using the actual media
URL in place of the mediaKey.

Note: A separate collection for game is unnecessary, so I’m using the
existing project collection with a portfolio property instead.